to handle c,s,%

+++ handle_cs%.c
@@ -37,7 +37,7 @@ int _printf(const char *format, ...)
 				write(1, format, 1);
 				break;
 			}
-			m = (*(format + 1) != 's') ? m + 1 : m + strlen(j);
+			m = (*(format + 1) != 's') ? m + 1 : m + (int)strlen(j);
 			format = (n == 1) ? format + 1 : format + 2;
 		}
 		else
